feat: blocking is added according to permissions, in views (@can) and controllers (Get)

+++ app/Http/Controllers/PermissionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 
 class PermissionController extends Controller
@@ -16,6 +17,8 @@ class PermissionController extends Controller
     public function index()
     {
         //
+        abort_if(Gate::denies('permission_index'), 403);
+
         $permissions = Permission::paginate(5);
 
         return view('permissions.index', compact('permissions'));
@@ -29,6 +32,8 @@ class PermissionController extends Controller
     public function create()
     {
         //
+        abort_if(Gate::denies('permission_create'), 403);
+
         return view('permissions.create');
     }
 
@@ -55,6 +60,8 @@ class PermissionController extends Controller
     public function show(Permission $permission)
     {
         //
+        abort_if(Gate::denies('permission_show'), 403);
+
         return view('permissions.show', compact('permission'));
 
     }
@@ -68,6 +75,8 @@ class PermissionController extends Controller
     public function edit(Permission $permission)
     {
         //
+        abort_if(Gate::denies('permission_edit'), 403);
+
         return view('permissions.edit', compact('permission'));
     }
 
@@ -95,6 +104,8 @@ class PermissionController extends Controller
     public function destroy(Permission $permission)
     {
         //
+        abort_if(Gate::denies('permission_delete'), 403);
+
         $permission->delete();
 
         return redirect()->route('permissions.index');
