+++ app/Http/Controllers/RoleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 use Symfony\Component\Console\Input\Input;
@@ -17,6 +18,8 @@ class RoleController extends Controller
     public function index()
     {
         //
+        abort_if(Gate::denies('role_index'), 403);
+
         $roles = Role::paginate(10);
 
         return view('roles.index', compact('roles'));
@@ -30,6 +33,8 @@ class RoleController extends Controller
     public function create()
     {
         //
+        abort_if(Gate::denies('role_create'), 403);
+
         $permissions = Permission::all()->pluck('name','id');
         // dd($permissions);
         return view('roles.create', compact('permissions'));
@@ -63,6 +68,8 @@ class RoleController extends Controller
         //
         // abort_if(Gate::denies('role_show'), 403);
 
+        abort_if(Gate::denies('role_show'), 403);
+
         $role->load('permissions');
         return view('roles.show', compact('role'));
     }
@@ -76,6 +83,8 @@ class RoleController extends Controller
     public function edit(Role $role)
     {
         //
+        abort_if(Gate::denies('role_edit'), 403);
+
         $permissions = Permission::all()->pluck('name','id');
 
         $role->load('permissions');
@@ -95,7 +104,7 @@ class RoleController extends Controller
     {
         //
         $role->update($request->only('name'));
-        
+
         // $role->permissions()->sync($request->input('permissions', []));
         $role->syncPermissions($request->input('permissions', []));
 
@@ -111,6 +120,8 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         //
+        abort_if(Gate::denies('role_delete'), 403);
+
         $role->delete();
 
         return redirect()->route('roles.index');
