+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Role;
 use App\Http\Requests\UserEditRequest;
 use App\Http\Requests\UserCreateRequest;
@@ -13,12 +14,14 @@ class UserController extends Controller
     //
     public function index()
     {
+        abort_if(Gate::denies('user_index'), 403);
         $users = User::paginate(5);
         return view('users.index', compact('users'));
     }
 
     public function create()
     {
+        abort_if(Gate::denies('user_create'), 403);
         $roles = Role::all()->pluck('name', 'id');
         return view('users.create', compact('roles'));
     }
@@ -32,7 +35,7 @@ class UserController extends Controller
         //     'password' => 'required',
         // ]);
 
-        $user = User::create($request->only('name', 'username', 'email') 
+        $user = User::create($request->only('name', 'username', 'email')
             + ['password' => bcrypt($request->input('password')),]);
 
         $roles = $request->input('roles', []);
@@ -43,6 +46,7 @@ class UserController extends Controller
 
     public function show(User $user)
     {
+        abort_if(Gate::denies('user_show'), 403);
         // $user=User::findOrFail($id);
         // dd($user);
         $user->load('roles');
@@ -51,6 +55,7 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
+        abort_if(Gate::denies('user_edit'), 403);
         $roles = Role::all()->pluck('name', 'id');
         $user->load('roles');
         return view('users.edit', compact('user', 'roles'));
@@ -63,7 +68,7 @@ class UserController extends Controller
 
         if($password)
             $data['password'] = bcrypt($password);
-        
+
         $user->update($data);
 
         $roles = $request->input('roles', []);
@@ -74,10 +79,12 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        abort_if(Gate::denies('user_destroy'), 403);
+
         if (auth()->user()->id == $user->id) {
             return redirect()->route('users.index');
         }
-        
+
         $user->delete();
 
         return back()->with('success', 'Successfully deleted user');
